Create Astermodel.py
This program trains a yolov8 model on large weights (pre-trained weights) so that it can be used in a computer vision program.

The author of this program is David Minott, debugging was done with AI. 

The import section of the code is broken down as follows:

Import section:
"os" - provides functions for us to interact with the operating system, in my program it is used to manage file paths

"PIL.Image" is the Pillow library (Python Image Library) that is used to handle image reading. We need this to be able to "read" our training and validation images. Image.open() will open an image file.

"torch" is the PyTorch library that is used for machine learning models. It creates the tools we need to handle tensors, creating NNs, and performing optimzations (This program is the first time I have used an optimizer-THE OPTIMZER IS NOT CUSTOM).

"torch.utils.data.Dataset" - is the PyTorch class for handling datasets (The datasets must follow a specific format, and have a .yaml file in them).

"torch.utils.DataLoader" - Loads the data from the dataset into batches (the batch size can change depending on how much memory you have) so that they can be used for training.

"torchvision.transforms"- Provides common image transformations such as skewing, color changes, and more. The image transform I performed in this program was a blue, because the movement of the robot may make the camera blurry.

"ultralytics.YOLO" - A class from the ultralytics package that intializes yolo models. In this case I used a yolov8 model as their is more documentation on it then a new yolov9 model. 

---------------------------------------------------------------

MORE SECTIONS WILL BE BROKEN DOWN IN THE NEXT FEW COMMENTS.